requirements.txt
streamlit
pyotp
qrcode
watchdog
bycrypt 
cryptography
db-sqlite3
python-dotenv
